+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Indexer Plugin ChangeLog
 
+## v1.0.2
+
+  - Improved error checking on getItemInfo()
+  - Improved indexing on utf-8 content
+  - New Czech translation
+  
 ## v1.0.1
   - Resolved an issue where saving an article did not properly re-index the content
   - PHP v7.4 fixes